firebase_functions/requirements.txt
flask>=2.0.0
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
functions-framework>=3.0.0 